Reduce module sprawl
Closely associated types can be stored in the same module. This
reduces the amount of re-exporting that has to happen to get
everything out, reduces file switching during development, and
may decrease the chance of creating confusing type name conflicts.
